refactor(completion): move subcommand lists to router.ts
TOP_LEVEL_SUBCOMMANDS and MARKETPLACE_SUBCOMMANDS now live in router.ts
alongside the switch statements they must mirror. provider.ts imports
them. Adding a new command to the router without updating the list is
now an obvious omission in the same file.
