+++ extensions/pi-claude-marketplace/edge/router.ts
@@ -42,6 +42,35 @@ export interface SubcommandHandlers {
   marketplaceNoautoupdate: (args: string, ctx: ExtensionCommandContext) => Promise<void>;
 }
 
+/**
+ * All top-level subcommands accepted by routeClaudePlugin, including
+ * aliases. Imported by the completion provider so both systems stay in sync.
+ */
+export const TOP_LEVEL_SUBCOMMANDS = [
+  "bootstrap",
+  "install",
+  "uninstall",
+  "update",
+  "list",
+  "ls",
+  "marketplace",
+] as const;
+
+/**
+ * All subcommands accepted by routeMarketplace, including aliases.
+ * Imported by the completion provider so both systems stay in sync.
+ */
+export const MARKETPLACE_SUBCOMMANDS = [
+  "add",
+  "remove",
+  "rm",
+  "list",
+  "ls",
+  "update",
+  "autoupdate",
+  "noautoupdate",
+] as const;
+
 export const TOP_LEVEL_USAGE =
   "Usage: /claude:plugin <bootstrap|install|uninstall|update|list|ls|marketplace> ...\n" +
   "  bootstrap                                          add anthropics/claude-plugins-official to user scope and enable autoupdate\n" +
